requirements.txt
numpy
matplotlib
jupyter
pillow
scikit-learn
scikit-image
scipy
h5py
eventlet
flask-socketio
seaborn
pandas
ffmpeg
imageio==2.1.2
moviepy
opencv-python
requests
keras==2.0.9
